// SPQuickEditList/src/models/IActionPlanItem.ts
export interface IActionPlanItem {
  brigadeName?: string;
  endState?: string;
  viabilityCategory?: string;
  subCategory?: string;
  rating?: string;
  statementSelection?: string;
  treatment?: string;
  initiative?: string;
  supportRequired?: string;
  priority?: string;
  due?: string;
  status?: string;
}